JUnit test has been added and passed.

+++ OOPDevelopmentVol2/src/info/sjd/AbstractShape.java
@@ -2,7 +2,7 @@ package info.sjd;
 
 public abstract class AbstractShape implements Shape {
 	
-	public abstract float getArea();
+	public abstract double getArea();
 
 
 }
