+++ OOPDevelopmentVol2/src/info/sjd/Shape.java
@@ -2,5 +2,5 @@ package info.sjd;
 
 public interface Shape {
 
-	float getArea();
+	double getArea();
 }
